+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Extract serverless utilities that are not specific to the AWS Lambda platform into their own utility package, @instana/serverless.
+- Log a warning when @instana/collector has been initialized too late. Additionally, this will be transmitted as snapshot data.
 
 ## 1.75.0
 - Capture HTTP response headers for HTTP entry spans.